Refactor termination queue logic
