+++ DSL/Ruuter.public/DSL/POST/add-chat-to-termination-queue.yml
@@ -1,9 +1,9 @@
 extractRequestData:
   assign:
-    chatId: "${incoming.body.message.chatId}"
-    event: "${incoming.body.message.event}"
-    authorRole: ${incoming.body.message.authorRole}
-    authorTimestamp: "${incoming.body.message.authorTimestamp}"
+    chatId: "${incoming.body.chatId}"
+    event: "${incoming.body.event}"
+    authorRole: "${incoming.body.authorRole}"
+    authorTimestamp: "${incoming.body.authorTimestamp}"
   next: updateChatStatus
 
 updateChatStatus:
@@ -14,7 +14,6 @@ updateChatStatus:
       chatId: ${chatId}
       scheduledForTerminated: ${new Date().toISOString()}
   result: statusUpdateRes
-  next: addChatToTerminationQueue
 
 addChatToTerminationQueue:
   call: http.post
@@ -26,7 +25,6 @@ addChatToTerminationQueue:
       authorRole: ${authorRole}
       authorTimestamp: ${authorTimestamp}
   result: res
-  next: return_chat_added_to_termination_queue
 
 return_chat_added_to_termination_queue:
   return: "Chat has added to termination queue"
